Fix the email regexp CPU usage for odd inputs
Fixes #34

Some inputs could make the Regexp engine in Ruby to take up a lot of
CPU time in order to check for a match of the email string. This change
adds a test to guard against an example of such an input and changes
the email regexp so that all the tests pass and so that no excessive
CPU usage occurs.