Include trailing hyphens in URL
Before this commit

http://www.rubyonrails.com/foo.cgi?trailing_ampersand=value-

Would generate

<a href=\"http://www.rubyonrails.com/foo.cgi?trailing_ampersand=value\">http://www.rubyonrails.com/foo.cgi?trailing_ampersand=value</a>-

Instead of

<a href=\"http://www.rubyonrails.com/foo.cgi?trailing_ampersand=value-\”>http://www.rubyonrails.com/foo.cgi?trailing_ampersand=value-</a>

The reason this was failing was because the unescaped `-` acted as a
character class (specifying a range of characters to match, rather than
matching against the literal hyphen character. The previously matched
characters were:

= & ; : / <

`=` and `&` already had specific test cases.

`test_autolink_with_trailing_amp_on_link` was failing when we escape the
hyphen (because `&` is turned into `&amp;` and because of that we rely
on `;` being included in the URL.

`/` was already explicitly included in the pattern but was not
explicitly tested.

`>` is not a problem as it’s escaped to `&lt;`.

Including `:` has been
[identified](https://github.com/tenderlove/rails_autolink/pull/45#issuecomment-1209691944)
as a regression so we add a test to document that a trailing `:` is no
longer being part of the URL.