Require a word part after “www.” when auto linking
This is to not parse things like “awww... what a
cute kitten” as a link.

Credits to @ItsYou who proposed the solution to
issue #46 back in 2015. This commit slims down the
tests a bit, leaves out the unnecessary forward-
slash escape and add regression test for numeric
domain but is otherwise the same as PR #58.